Fix crash if SpecialFolder.ApplicationData doesn't exist
If the folder doesn't exist, the config variable will be 'ILSpy.xml',
and the parent directory name will be an empty string. This will cause
CreateDirectory to throw, and the application to potentially crash due
to an exception in a background thread. This manifests as a native
crash, NSInternalInconsistencyException, on my machine, with the message
'Terminating app due to uncaught exception
'NSInternalInconsistencyException',reason: 'NSWindow drag regions should
only be invalidated on the Main Thread'.

By passing 'DoNotVerify' to Environment.GetFolderPath we can avoid the
check for directory existence and avoid the crash.